tools: improve error detection in find-inactive-collaborators
Previously, if the script failed to find a collaborator section in the
README file, it would carry on with no results and no error. This
detects the problem and throws an error.

It is also more robust in that it will still work if the emeriti section
ends up getting moved and does not immedaitely follow the collaborator
section.

PR-URL: https://github.com/nodejs/node/pull/39617
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Benjamin Gruenbaum <benjamingr@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>