tools,meta: remove exclusions from AUTHORS
The update-authors tool omits Chromium team members. This was a decision
made years ago. The idea was that these folks contributed to V8 not to
Node.js, as they typically ended up in our code-base via cherry-picked
commits from V8.
Ref: https://github.com/nodejs/node/pull/232#discussion-diff-22412788

Now that there are over 3000 contributors identified in AUTHORS, some of
whom are bots that we run, I think it's time to undo this exception. If
we used someone else's code in a cherry-pick, congratulations to them,
they're authors. In particular, this will fix the omission of Yang
Guo who contributed extensively to Node.js outside of V8 contributions.

PR-URL: https://github.com/nodejs/node/pull/40648
Reviewed-By: Antoine du Hamel <duhamelantoine1995@gmail.com>
Reviewed-By: Michaël Zasso <targos@protonmail.com>