doc: document flow for supporting type generation
Refs: https://github.com/nodejs/next-10/issues/104

Add initial doc to capture how the project supports
the flow for types generation. Based on discussions
from the Next-10 mini-summit that included project
members and members of the Definitely Typed team.

Will need to be updated as we improve the flow and
resulting json but captures the current flow and
approach and will give us the place where we can
update as things change.

Signed-off-by: Michael Dawson <mdawson@devrus.com>

PR-URL: https://github.com/nodejs/node/pull/41464
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Chengzhong Wu <legendecas@gmail.com>
Reviewed-By: Michaël Zasso <targos@protonmail.com>