build: drop shortened URL from lint-commit-message
The shortened link to the commit message guidelines no longer works
after they were moved from `docs/guides` to `docs/contributing`. Now
that `core-validate-commit` outputs an error message pointing to the
full URL of the commit message guidelines on failure we no longer need
to include the URL (shortened or otherwise) in the workflow title.

PR-URL: https://github.com/nodejs/node/pull/42168
Refs: https://github.com/nodejs/core-validate-commit/pull/95
Refs: https://github.com/nodejs/node/issues/41697
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: Michaël Zasso <targos@protonmail.com>
Reviewed-By: Mestery <mestery@protonmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Mohammed Keyvanzadeh <mohammadkeyvanzade94@gmail.com>
Reviewed-By: Darshan Sen <raisinten@gmail.com>