doc: update AIX minimum supported arch
Raise the minimum supported arch for AIX to Power 8 in line with the
existing minimum supported Power arch for Linux. We do not have anything
earlier than Power 8 in our CI so are unable to test on it, and have
reached out to IBM who have no concerns over baselining on Power 8 for
Node.js 18.

PR-URL: https://github.com/nodejs/node/pull/42604
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: Gireesh Punathil <gpunathi@in.ibm.com>
Reviewed-By: Michael Dawson <midawson@redhat.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Akhil Marsonya <akhil.marsonya27@gmail.com>
Reviewed-By: Darshan Sen <raisinten@gmail.com>