doc: update minimum glibc requirements for Linux
Starting with Node.js 18 we have moved building the release binaries
from being built on CentOS 7/RHEL 7 to being built on RHEL 8 as the
End-of-Life date of CentOS 7/RHEL 7 is before the End-of-Life date
of Node.js 18 and we do not want to have to make a disruptive change
towards the end of Node.js 18's life cycle.

Moving to RHEL 8 has raised the glibc version being linked against
(2.28). The official Node.js Linux release binaries will only run on
Linux distributions with a matching or higher version of glibc.

PR-URL: https://github.com/nodejs/node/pull/42659
Refs: https://github.com/nodejs/build/issues/2815
Refs: https://github.com/nodejs/build/issues/2741
Reviewed-By: Michael Dawson <midawson@redhat.com>
Reviewed-By: Stewart X Addison <sxa@redhat.com>
Reviewed-By: Beth Griggs <bgriggs@redhat.com>
Reviewed-By: Michaël Zasso <targos@protonmail.com>